+++ a1/exceptions.py
@@ -29,15 +29,3 @@ class PolicyTypeNotFound(BaseException):
 
 class PolicyTypeAlreadyExists(BaseException):
     """a policy type already exists and replace not supported at this time"""
-
-
-class MissingRmrString(BaseException):
-    pass
-
-
-class MissingManifest(BaseException):
-    pass
-
-
-class MissingRmrMapping(BaseException):
-    pass